+++ envoy-mobile.tulsiproj/Configs/all.tulsigen
@@ -11,8 +11,20 @@
     "//library/swift:ios_framework_archive",
     "//test/swift:counter_tests",
     "//test/swift:counter_tests_lib",
-    "//test/swift:direct_response_tests",
-    "//test/swift:direct_response_tests_lib",
+    "//test/swift:direct_response_contains_headers_integration_tests",
+    "//test/swift:direct_response_contains_headers_integration_tests_lib",
+    "//test/swift:direct_response_exact_headers_match_integration_tests",
+    "//test/swift:direct_response_exact_headers_match_integration_tests_lib",
+    "//test/swift:direct_response_exact_path_match_integration_tests",
+    "//test/swift:direct_response_exact_path_match_integration_tests_lib",
+    "//test/swift:direct_response_filter_mutation_integration_tests",
+    "//test/swift:direct_response_filter_mutation_integration_tests_lib",
+    "//test/swift:direct_response_prefix_headers_match_integration_tests",
+    "//test/swift:direct_response_prefix_headers_match_integration_tests_lib",
+    "//test/swift:direct_response_prefix_path_match_integration_tests",
+    "//test/swift:direct_response_prefix_path_match_integration_tests_lib",
+    "//test/swift:direct_response_suffix_path_match_integration_tests",
+    "//test/swift:direct_response_suffix_path_match_integration_tests_lib",
     "//test/swift:element_tests",
     "//test/swift:element_tests_lib",
     "//test/swift:engine_builder_tests",
